NF/BF/RF: Major update to allow for dask and joblib
This represents a significant refactor of the code to support
sending hierarchical jobs to dask and joblib with a functional
approach that avoids sending large variables to other processes.

The new code can support significant

Other smaller changes (should have been in a separate commit, but
this is where it happened) include conversion of internal dist calls
from pdf (followed by a log) to logpdf, which can both speed up
the calculations and occasionally work better (in the case of a normal
prior).

This change also includes removal of the option to calculate posteriors
during the sampling phase. It took up room and was a waste of computation
when the acceptance rates are low.