NF/BF/RF: Correct fixed param transforms, DE mods, delayed init for fixed, log acceptance.
Finalizes conversion to staying in log whenever possible. Here we now
stay in log space for the acceptance test in the M-H step.

The DE objects are now created in init for each module, determining
number of params to optimize gamma.

Fixes a transform bug that was using the incorrect transforms for FixedParams.

Now allows for delaying fixed param iterations during burnin if they modulate
a HyperPrior instance.